Changes long lat attribute to float data type

+++ db/migrate/20170126135435_add_lng_and_lat_to_users.rb
@@ -1,6 +1,6 @@
 class AddLngAndLatToUsers < ActiveRecord::Migration[5.0]
   def change
-    add_column :users, :lat, :integer
-    add_column :users, :lng, :integer
+    add_column :users, :lat, :float
+    add_column :users, :lng, :float
   end
 end
